chore(iast): fix import chains
